tools/openai-whisper/pyproject.toml
[project]
name = "openai-whisper-cli"
version = "0.1.0"
requires-python = ">=3.10,<3.14"
dependencies = [
  "openai-whisper",
  "torch",
  "torchaudio",
]
